+++ Source/WebKit/GPUProcess/graphics/RemoteGraphicsContextGLFunctionsGenerated.h
@@ -1382,4 +1382,11 @@
         m_context->getInternalformativ(target, internalformat, pname, params);
         completionHandler(IPC::ArrayReference<int32_t>(reinterpret_cast<int32_t*>(params.data()), params.size()));
     }
+    void paintRenderingResultsToPixelBuffer(CompletionHandler<void(RefPtr<WebCore::PixelBuffer>&&)>&& completionHandler)
+    {
+        RefPtr<WebCore::PixelBuffer> returnValue = { };
+        assertIsCurrent(workQueue());
+        returnValue = m_context->paintRenderingResultsToPixelBuffer();
+        completionHandler(WTFMove(returnValue));
+    }
